ipc: page tables; Make page table API generic for use outside IPC core
Allow users outside of IPC to make use of compressed host page tables.
